Version 2.8.0: bash helpers and more.
 - adding circle CI (but not passing yet)
 - adding screenshots in a separate markdown
 - lots of fixes to bash helpers, in particular to sym.symit
 - adding sym-encrypt and sym-decrypt bash helpers
 - simplifying error reporting by Sym
 - updating CHANGELOG
 - ensuring sym exits with a non-zero code upon error

+++ lib/sym/app/commands/bash_completion.rb
@@ -7,26 +7,43 @@ module Sym
         required_options [:bash_support]
         try_after :generate_key, :open_editor, :encrypt, :decrypt
 
+        def ok
+          ' '.bold.green
+        end
+
         def execute
           file = opts[:bash_support]
 
           out = ''
           Sym::Constants::Bash::Config.each_pair do |key, config|
             script_name = key.to_s
-            FileUtils.cp(config[:source], config[:dest])
+
+            # This removes the old version of this file.
+            remove_old_version(out, config[:dest])
+
+            if File.exist?(config[:dest]) && File.read(config[:source]) == File.read(config[:dest])
+              out << "#{ok} file #{config[:dest].bold.blue } exists, and is up to date.\n"
+            else
+              FileUtils.cp(config[:source], config[:dest])
+              out << "#{ok} installing #{config[:dest].bold.blue }...\n"
+            end
+
             out << if File.exist?(file)
                      if File.read(file).include?(config[:script])
-                       "#{'OK'.bold.green}, #{file.bold.yellow} already has #{script_name.bold.blue} installed\n"
+                       "#{ok} BASH script #{file.bold.yellow} already sources #{script_name.bold.blue}.\n"
                      else
                        append_completion_script(file, config[:script])
-                       "#{'OK'.bold.green}, appended initialization for #{script_name.bold.blue} to #{file.bold.yellow}\n"
+                       "#{ok} BASH script #{script_name.bold.blue} is now sourced from #{file.bold.yellow}\n"
                      end
                    else
                      append_completion_script(file, config[:script])
-                     "#{'OK'.bold.green}, created new file #{file.bold.yellow}, added #{script_name.bold.blue} initialization.\n"
+                     "#{ok}, created new file #{file.bold.yellow}, added #{script_name.bold.blue} initialization.\n"
                    end
           end
-          out + "Please reload your terminal session to activate bash completion and other installed utilities.\n"
+          out << "\nPlease reload your terminal session to activate bash completion\n"
+          out << "and other installed BASH utilities.\n"
+          out << "\nAlternatively, just type #{"source #{file}".bold.green} to reload BASH.\n"
+          out << "Also — go ahead and try running #{"sym -h".bold.blue} and #{"symit -h".bold.blue}.\n"
         end
 
         private
@@ -37,6 +54,15 @@ module Sym
           end
         end
 
+        def remove_old_version(out, file)
+          if file =~ /\.bash$/
+            old_file = file.gsub(/\.bash$/, '')
+            if File.exist?(old_file)
+              out << "Removing old version — #{old_file.bold.magenta}..."
+              FileUtils.rm_f old_file
+            end
+          end
+        end
       end
     end
   end
